+++ Poseiden-skeleton/src/main/java/com/nnk/springboot/repositories/RuleRepository.java
@@ -4,5 +4,5 @@ import com.nnk.springboot.domain.Rule;
 import org.springframework.data.jpa.repository.JpaRepository;
 
 
-public interface RuleRepository extends JpaRepository<Rule, Integer> {
+public interface RuleRepository extends JpaRepository<Rule, Long> {
 }
